requirements.txt
dj-database-url==0.5.0
Django==2.1.4
django-admin-sortable==2.1.9
django-heroku==0.3.1
gunicorn==19.9.0
psycopg2==2.7.6.1
pytz==2018.7
whitenoise==4.1.2
